cleantup monitors
